refactor: split claude runtime helpers
