+++ connector/connector/runtimes/claude/runtime.py
@@ -1,18 +1,13 @@
 from __future__ import annotations
 
 import asyncio
-import hashlib
-import inspect
-import json
 import secrets
-from collections.abc import AsyncIterator, Callable, Mapping
+from collections.abc import Mapping
 from dataclasses import dataclass, field
 from typing import Any
 
 from connector.runtime_protocol import (
     AgentRuntime,
-    MarkdownMessageContent,
-    MessageTimelineItem,
     RuntimeAttachment,
     RuntimeCapability,
     RuntimeCapabilitySet,
@@ -22,27 +17,34 @@ from connector.runtime_protocol import (
     RuntimeOperationResult,
     RuntimePermissionCatalog,
     RuntimeSessionStateCache,
-    RuntimeTimelineItem,
     SessionMeta,
     SessionNotice,
     SessionState,
-    TimelineSource,
 )
 from connector.runtime_protocol.host import RuntimeHostClient
 from connector.runtimes.claude import provider_config
-
-SdkLoader = Callable[[], Any]
-ClaudeClientFactory = Callable[[Any, Any], Any]
-
-
-@dataclass(slots=True)
-class ClaudeSession:
-    session_id: str
-    external_session_id: str | None = None
-    cwd: str | None = None
-    active_turn_id: str | None = None
-    active_task: asyncio.Task[None] | None = None
-    client: Any = None
+from connector.runtimes.claude.domain.session import ClaudeSession
+from connector.runtimes.claude.sdk.client import (
+    ClaudeClientFactory,
+    SdkLoader,
+    connect_client,
+    disconnect_client,
+    interrupt_client,
+    load_sdk,
+    new_sdk_client,
+    query_client,
+    receive_response_messages,
+)
+from connector.runtimes.claude.timeline.messages import (
+    ClaudeMessageProjector,
+    is_result_message,
+    message_error_text,
+    message_id,
+    message_is_error,
+    message_role,
+    message_session_id,
+    message_text,
+)
 
 
 @dataclass(slots=True)
@@ -54,11 +56,11 @@ class ClaudeRuntime(AgentRuntime):
     runtime_version: str = "native-0"
     _sessions: dict[str, ClaudeSession] = field(default_factory=dict, init=False)
     _session_states: RuntimeSessionStateCache = field(init=False)
-    _order_by_id: dict[str, int] = field(default_factory=dict, init=False)
-    _next_order_seq: int = field(default=1, init=False)
+    _timeline: ClaudeMessageProjector = field(init=False)
 
     def __post_init__(self) -> None:
         self._session_states = RuntimeSessionStateCache("claude", self.host)
+        self._timeline = ClaudeMessageProjector()
 
     @property
     def identity(self) -> RuntimeIdentity:
@@ -77,7 +79,7 @@ class ClaudeRuntime(AgentRuntime):
             if session.active_task is not None and not session.active_task.done():
                 session.active_task.cancel()
                 tasks.append(session.active_task)
-            await _disconnect_client(session.client)
+            await disconnect_client(session.client)
             session.client = None
             session.active_turn_id = None
         if tasks:
@@ -320,11 +322,7 @@ class ClaudeRuntime(AgentRuntime):
                 message="Claude runtime has no active turn to interrupt",
             )
 
-        interrupted = False
-        interrupt = getattr(session.client, "interrupt", None)
-        if callable(interrupt):
-            await _maybe_await(interrupt())
-            interrupted = True
+        interrupted = await interrupt_client(session.client)
         if session.active_task is not None and not session.active_task.done():
             session.active_task.cancel()
             interrupted = True
@@ -356,19 +354,22 @@ class ClaudeRuntime(AgentRuntime):
         client_message_id: str | None,
     ) -> None:
         try:
-            sdk = self._load_sdk()
-            client = self._new_client(sdk, session)
+            sdk = load_sdk(self.sdk_loader)
+            client = new_sdk_client(
+                sdk=sdk,
+                config_values=self.config.values,
+                session=session,
+                client_factory=self.client_factory,
+            )
             session.client = client
-            connect = getattr(client, "connect", None)
-            if callable(connect):
-                await _maybe_await(connect())
+            await connect_client(client)
             await self._set_session_state(
                 session,
                 "running",
                 metadata={"source": "claude.turn.running", "turnId": turn_id},
             )
             await self.host.timeline_item_upsert(
-                self._message_item(
+                self._timeline.message_item(
                     session=session,
                     turn_id=turn_id,
                     role="user",
@@ -377,37 +378,34 @@ class ClaudeRuntime(AgentRuntime):
                     client_message_id=client_message_id,
                 )
             )
-            query = getattr(client, "query", None)
-            if not callable(query):
-                raise RuntimeError("ClaudeSDKClient.query is unavailable")
-            await _maybe_await(query(content))
+            await query_client(client, content)
 
             result_error: Mapping[str, Any] | None = None
-            async for message in _receive_response_messages(client):
-                external_session_id = _message_session_id(message)
+            async for message in receive_response_messages(client):
+                external_session_id = message_session_id(message)
                 if external_session_id is not None:
                     session.external_session_id = external_session_id
-                if _is_result_message(message):
-                    if _message_is_error(message):
+                if is_result_message(message):
+                    if message_is_error(message):
                         result_error = {
                             "code": "claude_result_error",
-                            "message": _message_error_text(message)
+                            "message": message_error_text(message)
                             or "Claude turn completed with an error",
                         }
                     continue
-                if _message_role(message) != "assistant":
+                if message_role(message) != "assistant":
                     continue
-                text = _message_text(message)
+                text = message_text(message)
                 if not text:
                     continue
                 await self.host.timeline_item_upsert(
-                    self._message_item(
+                    self._timeline.message_item(
                         session=session,
                         turn_id=turn_id,
                         role="assistant",
                         text=text,
                         event="claude.turn.assistant",
-                        native_item_id=_message_id(message),
+                        native_item_id=message_id(message),
                     )
                 )
 
@@ -441,47 +439,10 @@ class ClaudeRuntime(AgentRuntime):
                 metadata={"source": "claude.turn.failed", "turnId": turn_id},
             )
         finally:
-            await _disconnect_client(session.client)
+            await disconnect_client(session.client)
             session.client = None
             if session.active_turn_id == turn_id:
                 session.active_turn_id = None
-
-    def _load_sdk(self) -> Any:
-        if self.sdk_loader is None:
-            raise RuntimeError("Claude SDK loader is not configured")
-        return self.sdk_loader()
-
-    def _new_client(self, sdk: Any, session: ClaudeSession) -> Any:
-        options = self._sdk_options(sdk, session)
-        if self.client_factory is not None:
-            return self.client_factory(sdk, options)
-        client_cls = getattr(sdk, "ClaudeSDKClient", None)
-        if client_cls is None:
-            raise RuntimeError("ClaudeSDKClient is not available")
-        try:
-            return client_cls(options=options)
-        except TypeError:
-            return client_cls(options)
-
-    def _sdk_options(self, sdk: Any, session: ClaudeSession) -> Any:
-        values = dict(self.config.values)
-        kwargs: dict[str, Any] = {"include_partial_messages": True}
-        if session.cwd:
-            kwargs["cwd"] = session.cwd
-        if session.external_session_id:
-            kwargs["resume"] = session.external_session_id
-        executable_path = values.get("executablePath")
-        if isinstance(executable_path, str) and executable_path:
-            kwargs["cli_path"] = executable_path
-        environment = values.get("environment")
-        if isinstance(environment, Mapping):
-            kwargs["env"] = dict(environment)
-        options_cls = getattr(sdk, "ClaudeAgentOptions", None) or getattr(
-            sdk, "ClaudeCodeOptions", None
-        )
-        if options_cls is None:
-            return kwargs
-        return options_cls(**kwargs)
 
     def _session_for(
         self,
@@ -518,47 +479,6 @@ class ClaudeRuntime(AgentRuntime):
             metadata=metadata,
         )
 
-    def _message_item(
-        self,
-        session: ClaudeSession,
-        turn_id: str,
-        role: str,
-        text: str,
-        event: str,
-        client_message_id: str | None = None,
-        native_item_id: str | None = None,
-    ) -> RuntimeTimelineItem:
-        stable_key = native_item_id or client_message_id or text
-        item_id = _stable_id(
-            "message",
-            session.session_id,
-            session.external_session_id,
-            turn_id,
-            role,
-            stable_key,
-        )
-        order_seq = self._order_by_id.get(item_id)
-        if order_seq is None:
-            order_seq = self._next_order_seq
-            self._next_order_seq += 1
-            self._order_by_id[item_id] = order_seq
-        return MessageTimelineItem(
-            id=item_id,
-            type="message",
-            status="done",
-            role=role,  # type: ignore[arg-type]
-            turn_id=turn_id,
-            content=MarkdownMessageContent(text=text),
-            source=TimelineSource(
-                runtime="claude",
-                external_session_id=session.external_session_id,
-                turn_id=turn_id,
-                native_item_id=native_item_id,
-                event=event,
-                client_message_id=client_message_id,
-            ),
-        ).to_platform_item(session_id=session.session_id, order_seq=order_seq)
-
     def _has_active_turn(self, session_id: str) -> bool:
         session = self._sessions.get(session_id)
         return bool(
@@ -566,123 +486,3 @@ class ClaudeRuntime(AgentRuntime):
             and session.active_task is not None
             and not session.active_task.done()
         )
-
-
-async def _receive_response_messages(client: Any) -> AsyncIterator[Any]:
-    receive_response = getattr(client, "receive_response", None)
-    if not callable(receive_response):
-        return
-    response = receive_response()
-    if hasattr(response, "__aiter__"):
-        async for message in response:
-            yield message
-        return
-    messages = await _maybe_await(response)
-    if messages is None:
-        return
-    for message in messages:
-        yield message
-
-
-async def _disconnect_client(client: Any) -> None:
-    disconnect = getattr(client, "disconnect", None)
-    if callable(disconnect):
-        await _maybe_await(disconnect())
-
-
-async def _maybe_await(value: Any) -> Any:
-    if inspect.isawaitable(value):
-        return await value
-    return value
-
-
-def _message_role(message: Any) -> str | None:
-    raw_role = _extract(message, "role")
-    if isinstance(raw_role, str) and raw_role:
-        return raw_role
-    nested = _extract(message, "message")
-    if isinstance(nested, Mapping):
-        raw_nested_role = nested.get("role")
-        if isinstance(raw_nested_role, str) and raw_nested_role:
-            return raw_nested_role
-    raw_type = _extract(message, "type")
-    return raw_type if isinstance(raw_type, str) and raw_type else None
-
-
-def _message_text(message: Any) -> str | None:
-    nested = _extract(message, "message")
-    if isinstance(nested, Mapping):
-        text = _content_text(nested.get("content"))
-        if text:
-            return text
-    text = _content_text(_extract(message, "content"))
-    if text:
-        return text
-    result = _extract(message, "result")
-    return result if isinstance(result, str) and result else None
-
-
-def _content_text(content: Any) -> str | None:
-    if isinstance(content, str):
-        return content
-    if not isinstance(content, list | tuple):
-        return None
-    parts: list[str] = []
-    for block in content:
-        text = _extract(block, "text")
-        if isinstance(text, str) and text:
-            parts.append(text)
-    return "\n".join(parts) if parts else None
-
-
-def _message_session_id(message: Any) -> str | None:
-    value = _extract(message, "session_id", "sessionId")
-    return value if isinstance(value, str) and value else None
-
-
-def _message_id(message: Any) -> str | None:
-    value = _extract(message, "uuid", "message_id", "messageId", "id")
-    return value if isinstance(value, str) and value else None
-
-
-def _is_result_message(message: Any) -> bool:
-    if message.__class__.__name__ == "ResultMessage":
-        return True
-    raw_type = _extract(message, "type")
-    subtype = _extract(message, "subtype")
-    return raw_type == "result" or isinstance(subtype, str) and "result" in subtype
-
-
-def _message_is_error(message: Any) -> bool:
-    return _extract(message, "is_error", "isError") is True
-
-
-def _message_error_text(message: Any) -> str | None:
-    errors = _extract(message, "errors")
-    if isinstance(errors, list) and errors:
-        return "; ".join(str(error) for error in errors)
-    value = _extract(message, "error", "terminal_reason", "terminalReason")
-    return value if isinstance(value, str) and value else None
-
-
-def _extract(value: Any, *names: str) -> Any:
-    if isinstance(value, Mapping):
-        for name in names:
-            if name in value:
-                return value[name]
-        return None
-    for name in names:
-        candidate = getattr(value, name, None)
-        if candidate is not None:
-            return candidate
-    return None
-
-
-def _stable_id(*parts: Any) -> str:
-    payload = json.dumps(
-        parts,
-        ensure_ascii=False,
-        sort_keys=True,
-        separators=(",", ":"),
-    )
-    return "claude_" + hashlib.sha256(payload.encode("utf-8")).hexdigest()[:24]
