+++ connector/connector/runtimes/claude/sdk/__init__.py
@@ -1,0 +1,1 @@
+"""Claude Agent SDK integration helpers."""
