+++ connector/connector/runtimes/claude/timeline/__init__.py
@@ -1,0 +1,1 @@
+"""Claude timeline projection helpers."""
